Move FullNetwork trait to reth-network-api
