+++ crates/node/api/Cargo.toml
@@ -17,7 +17,7 @@ reth-provider.workspace = true
 reth-db-api.workspace = true
 reth-engine-primitives.workspace = true
 reth-transaction-pool.workspace = true
-reth-network.workspace = true
 reth-payload-builder.workspace = true
 reth-payload-primitives.workspace = true
 reth-tasks.workspace = true
+reth-network-api.workspace = true
