+++ crates/node/builder/Cargo.toml
@@ -49,6 +49,7 @@ reth-rpc-types.workspace = true
 reth-engine-util.workspace = true
 reth-cli-util.workspace = true
 reth-rpc-eth-types.workspace = true
+reth-network-api.workspace = true
 
 ## async
 futures.workspace = true
